Refactor
Revised file structure
